Create NodeProvider
