+++ randovania/game_description/data_writer.py
@@ -15,8 +15,8 @@ from randovania.game_description.resources.simple_resource_info import SimpleRes
 from randovania.game_description.resources.trick_resource_info import TrickResourceInfo
 from randovania.game_description.world.area import Area
 from randovania.game_description.world.dock import DockWeaknessDatabase, DockWeakness
-from randovania.game_description.world.node import Node, GenericNode, DockNode, PickupNode, TeleporterNode, EventNode, \
-    ConfigurableNode, LogbookNode, LoreType, PlayerShipNode
+from randovania.game_description.world.node import Node, GenericNode, DockNode, TeleporterNode, ConfigurableNode
+from randovania.game_description.world.resource_node import PickupNode, EventNode, LoreType, LogbookNode, PlayerShipNode
 from randovania.game_description.world.world import World
 from randovania.game_description.world.world_list import WorldList
 from randovania.lib import frozen_lib
