+++ randovania/game_description/world/area.py
@@ -4,7 +4,8 @@ from typing import List, Dict, Optional, Iterator, Tuple
 
 from randovania.game_description.requirements import Requirement
 from randovania.game_description.resources.pickup_index import PickupIndex
-from randovania.game_description.world.node import Node, PickupNode
+from randovania.game_description.world.node import Node
+from randovania.game_description.world.resource_node import PickupNode
 
 
 @dataclasses.dataclass(frozen=True)
