+++ randovania/games/binary_data.py
@@ -7,7 +7,7 @@ from construct import (Struct, Int32ub, Const, Byte, Float32b, Flag,
                        Short, PrefixedArray, Switch, VarInt, Float64b, Compressed)
 
 from randovania.game_description import game_migration
-from randovania.game_description.world.node import LoreType
+from randovania.game_description.world.resource_node import LoreType
 from randovania.games.game import RandovaniaGame
 from randovania.lib.construct_lib import String, convert_to_raw_python, OptionalValue, ConstructDict, JsonEncodedValue
