+++ randovania/games/prime2/gui/hint_details_tab.py
@@ -6,7 +6,7 @@ from PySide6 import QtWidgets
 from randovania.exporter.hints.hint_exporter import HintExporter
 from randovania.game_description import default_database
 from randovania.game_description.game_patches import GamePatches
-from randovania.game_description.world.node import LogbookNode
+from randovania.game_description.world.resource_node import LogbookNode
 from randovania.games.game import RandovaniaGame
 from randovania.games.prime2.exporter.hint_namer import EchoesHintNamer
 from randovania.gui.game_details.game_details_tab import GameDetailsTab
