+++ randovania/gui/dialog/node_details_popup.py
@@ -14,8 +14,9 @@ from randovania.game_description.resources.pickup_index import PickupIndex
 from randovania.game_description.resources.search import MissingResource, find_resource_info_with_long_name
 from randovania.game_description.world.area import Area
 from randovania.game_description.world.area_identifier import AreaIdentifier
-from randovania.game_description.world.node import Node, GenericNode, DockNode, PickupNode, TeleporterNode, EventNode, \
-    ConfigurableNode, LogbookNode, LoreType, NodeLocation, PlayerShipNode
+from randovania.game_description.world.node import Node, GenericNode, DockNode, TeleporterNode, ConfigurableNode, \
+    NodeLocation
+from randovania.game_description.world.resource_node import PickupNode, EventNode, LoreType, LogbookNode, PlayerShipNode
 from randovania.game_description.world.world import World
 from randovania.gui.dialog.connections_editor import ConnectionsEditor
 from randovania.gui.generated.node_details_popup_ui import Ui_NodeDetailsPopup
