+++ randovania/resolver/bootstrap.py
@@ -7,7 +7,8 @@ from randovania.game_description.game_patches import GamePatches
 from randovania.game_description.resources.resource_database import ResourceDatabase
 from randovania.game_description.resources.resource_info import CurrentResources, \
     add_resource_gain_to_current_resources
-from randovania.game_description.world.node import PlayerShipNode, NodeContext, ResourceNode
+from randovania.game_description.world.node import NodeContext
+from randovania.game_description.world.resource_node import PlayerShipNode, ResourceNode
 from randovania.layout.base.base_configuration import BaseConfiguration
 from randovania.layout.base.major_items_configuration import MajorItemsConfiguration
 from randovania.layout.base.trick_level import LayoutTrickLevel
@@ -82,9 +83,8 @@ class Bootstrap:
             assert isinstance(starting_node, ResourceNode)
             add_resource_gain_to_current_resources(
                 starting_node.resource_gain_on_collect(NodeContext(
-                    game.world_list.identifier_for_node(starting_node),
-                    patches, initial_resources, game.world_list.all_nodes,
-                    game.resource_database,
+                    patches, initial_resources,
+                    game.resource_database, game.world_list,
                 )),
                 initial_resources,
             )
