+++ randovania/resolver/resolver.py
@@ -5,7 +5,8 @@ from randovania.game_description import default_database
 from randovania.game_description.game_patches import GamePatches
 from randovania.game_description.requirements import RequirementSet, RequirementList
 from randovania.game_description.resources.resource_info import ResourceInfo
-from randovania.game_description.world.node import PickupNode, ResourceNode, EventNode, Node
+from randovania.game_description.world.node import Node
+from randovania.game_description.world.resource_node import PickupNode, EventNode, ResourceNode
 from randovania.layout.base.base_configuration import BaseConfiguration
 from randovania.resolver import debug, event_pickup
 from randovania.resolver.event_pickup import EventPickupNode
@@ -58,7 +59,7 @@ def _should_check_if_action_is_safe(state: State,
     :return:
     """
     if any(resource in dangerous_resources
-           for resource in action.resource_gain_on_collect(state.context_for(action))):
+           for resource in action.resource_gain_on_collect(state.context_for())):
         return False
 
     if isinstance(action, EventNode):
