+++ randovania/resolver/state.py
@@ -10,7 +10,8 @@ from randovania.game_description.resources.resource_database import ResourceData
 from randovania.game_description.resources.resource_info import ResourceInfo, CurrentResources, \
     add_resource_gain_to_current_resources, add_resources_into_another, convert_resource_gain_to_current_resources
 from randovania.game_description.resources.resource_type import ResourceType
-from randovania.game_description.world.node import ResourceNode, Node, NodeContext
+from randovania.game_description.world.node import Node, NodeContext
+from randovania.game_description.world.resource_node import ResourceNode
 from randovania.game_description.world.world_list import WorldList
 
 
@@ -124,12 +125,12 @@ class State:
         :return:
         """
 
-        if not node.can_collect(self.context_for(node)):
+        if not node.can_collect(self.context_for()):
             raise ValueError(
                 "Trying to collect an uncollectable node'{}'".format(node))
 
         new_resources = copy.copy(self.resources)
-        add_resource_gain_to_current_resources(node.resource_gain_on_collect(self.context_for(node)),
+        add_resource_gain_to_current_resources(node.resource_gain_on_collect(self.context_for()),
                                                new_resources)
 
         energy = new_energy
@@ -193,13 +194,12 @@ class State:
             self.game_data,
         )
 
-    def context_for(self, node: Node) -> NodeContext:
+    def context_for(self) -> NodeContext:
         return NodeContext(
-            self.game_data.world_list.identifier_for_node(node),
             self.patches,
             self.resources,
-            self.game_data.world_list.all_nodes,
             self.resource_database,
+            self.game_data.world_list,
         )
