+++ test/games/prime/patcher_file_lib/test_item_hints.py
@@ -14,7 +14,7 @@ from randovania.game_description.hint import (
 from randovania.game_description.resources.pickup_index import PickupIndex
 from randovania.game_description.world.area import Area
 from randovania.game_description.world.area_identifier import AreaIdentifier
-from randovania.game_description.world.node import LogbookNode, PickupNode
+from randovania.game_description.world.resource_node import PickupNode, LogbookNode
 from randovania.game_description.world.world import World
 from randovania.game_description.world.world_list import WorldList
 from randovania.games.game import RandovaniaGame
