+++ test/resolver/test_resolver_reach.py
@@ -1,6 +1,6 @@
 from unittest.mock import MagicMock, PropertyMock
 
-from randovania.game_description.world.node import EventNode
+from randovania.game_description.world.resource_node import EventNode
 from randovania.resolver.resolver_reach import ResolverReach
